requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
pydantic==2.4.2
python-dotenv==1.0.0
google-cloud-vision==3.4.5
google-cloud-translate==3.11.3
spacy==3.7.2
nltk==3.8.1
pandas==2.1.1
matplotlib==3.8.0
pillow==10.0.1
pytest==7.4.2
httpx==0.25.0
sqlalchemy==2.0.22
psycopg2-binary==2.9.9
pgvector==0.2.3
sentence-transformers==2.2.2
openai==1.3.7
numpy==1.26.0
psycopg==3.1.12
